Changed magic command for Kusto to kql in KustoKernelConnection and changed methods to be async. Consolidated MsKustoKernel and MsSqlKernel into base class ToolsServiceKernel
