+++ src/Microsoft.DotNet.Interactive.SqlServer/MsSqlKernel.cs
@@ -2,297 +2,46 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Generic;
 using System.CommandLine;
 using System.CommandLine.Invocation;
 using System.CommandLine.Parsing;
-using System.ComponentModel;
-using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
 
 using Microsoft.DotNet.Interactive.Commands;
-using Microsoft.DotNet.Interactive.Events;
-using Microsoft.DotNet.Interactive.ExtensionLab;
 using Microsoft.DotNet.Interactive.Formatting;
-using Microsoft.DotNet.Interactive.Formatting.TabularData;
 
 namespace Microsoft.DotNet.Interactive.SqlServer
 {
-    public class MsSqlKernel :
-        Kernel,
-        IKernelCommandHandler<SubmitCode>,
-        IKernelCommandHandler<RequestCompletions>
+    public class MsSqlKernel : ToolsServiceKernel
     {
-        private bool _connected;
-        private bool _intellisenseReady;
-        private readonly Uri _tempFileUri;
         private readonly string _connectionString;
-        private readonly MsSqlServiceClient _serviceClient;
-
-        private readonly TaskCompletionSource<ConnectionCompleteParams> _connectionCompleted = new();
-
-        private Func<QueryCompleteParams, Task> _queryCompletionHandler;
-        private Func<MessageParams, Task> _queryMessageHandler;
 
         public MsSqlKernel(
             string name,
             string connectionString,
-            MsSqlServiceClient client) : base(name)
+            MsSqlServiceClient client) : base(name, client)
         {
             if (string.IsNullOrWhiteSpace(connectionString))
             {
                 throw new ArgumentException("Value cannot be null or whitespace.", nameof(connectionString));
             }
-
-            var filePath = Path.GetTempFileName();
-            _tempFileUri = new Uri(filePath);
+            
             _connectionString = connectionString;
-
-            _serviceClient = client ?? throw new ArgumentNullException(nameof(client));
-            _serviceClient.Initialize();
-
-            _serviceClient.OnConnectionComplete += HandleConnectionComplete;
-            _serviceClient.OnQueryComplete += HandleQueryComplete;
-            _serviceClient.OnIntellisenseReady += HandleIntellisenseReady;
-            _serviceClient.OnQueryMessage += HandleQueryMessage;
-
-            RegisterForDisposal(() =>
-            {
-                if (_connected)
-                {
-                    Task.Run(() => _serviceClient.DisconnectAsync(_tempFileUri)).Wait();
-                }
-            });
-            RegisterForDisposal(() => File.Delete(_tempFileUri.LocalPath));
         }
 
-        private void HandleConnectionComplete(object sender, ConnectionCompleteParams connParams)
+        public override async Task ConnectAsync()
         {
-            if (connParams.OwnerUri.Equals(_tempFileUri.AbsolutePath))
+            if (!Connected)
             {
-                if (connParams.ErrorMessage is not null)
-                {
-                    _connectionCompleted.SetException(new Exception(connParams.ErrorMessage));
-                }
-                else
-                {
-                    _connectionCompleted.SetResult(connParams);
-                }
+                await ServiceClient.ConnectAsync(TempFileUri, _connectionString);
+                await ConnectionCompleted.Task;
+                Connected = true;
             }
         }
 
-        private void HandleQueryComplete(object sender, QueryCompleteParams queryParams)
+        protected override Type GetType(string typeName)
         {
-            if (_queryCompletionHandler is not null)
-            {
-                Task.Run(() => _queryCompletionHandler(queryParams)).Wait();
-            }
-        }
-
-        private void HandleQueryMessage(object sender, MessageParams messageParams)
-        {
-            if (_queryMessageHandler is not null)
-            {
-                Task.Run(() => _queryMessageHandler(messageParams)).Wait();
-            }
-        }
-
-        private void HandleIntellisenseReady(object sender, IntelliSenseReadyParams readyParams)
-        {
-            if (readyParams.OwnerUri.Equals(_tempFileUri.AbsolutePath))
-            {
-                _intellisenseReady = true;
-            }
-        }
-
-        public async Task ConnectAsync()
-        {
-            if (!_connected)
-            {
-                await _serviceClient.ConnectAsync(_tempFileUri, _connectionString);
-                await _connectionCompleted.Task;
-                _connected = true;
-            }
-        }
-
-        public async Task HandleAsync(SubmitCode command, KernelInvocationContext context)
-        {
-            if (!_connected)
-            {
-                return;
-            }
-
-            // If a query handler is already defined, then it means another query is already running in parallel.
-            // We only want to run one query at a time, so we display an error here instead.
-            if (_queryCompletionHandler is not null)
-            {
-                context.Display("Error: Another query is currently running. Please wait for that query to complete before re-running this cell.");
-                return;
-            }
-
-            var completion = new TaskCompletionSource<bool>();
-
-            _queryCompletionHandler = async queryParams =>
-            {
-                try
-                {
-                    foreach (var batchSummary in queryParams.BatchSummaries)
-                    {
-                        foreach (var resultSummary in batchSummary.ResultSetSummaries)
-                        {
-                            if (completion.Task.IsCompleted)
-                            {
-                                return;
-                            }
-
-                            if (resultSummary.RowCount > 0)
-                            {
-                                var subsetParams = new QueryExecuteSubsetParams
-                                {
-                                    OwnerUri = _tempFileUri.AbsolutePath,
-                                    BatchIndex = batchSummary.Id,
-                                    ResultSetIndex = resultSummary.Id,
-                                    RowsStartIndex = 0,
-                                    RowsCount = Convert.ToInt32(resultSummary.RowCount)
-                                };
-                                var subsetResult = await _serviceClient.ExecuteQueryExecuteSubsetAsync(subsetParams);
-                                var tables = GetEnumerableTables(resultSummary.ColumnInfo, subsetResult.ResultSubset.Rows);
-                                foreach (var table in tables)
-                                {
-                                    var explorer = new NteractDataExplorer(table.ToTabularDataResource());
-                                    context.Display(explorer);
-                                }
-                            }
-                            else
-                            {
-                                context.Display($"Info: No rows were returned for query {resultSummary.Id} in batch {batchSummary.Id}.");
-                            }
-                        }
-                    }
-
-                    completion.SetResult(true);
-                }
-                catch (Exception e)
-                {
-                    completion.SetException(e);
-                }
-            };
-
-#pragma warning disable 1998
-            _queryMessageHandler = async messageParams =>
-            {
-                try
-                {
-                    if (messageParams.Message.IsError)
-                    {
-                        context.Fail(command, message: messageParams.Message.Message);
-                        completion.SetResult(true);
-                    }
-                    else
-                    {
-                        context.Display(messageParams.Message.Message);
-                    }
-                }
-                catch (Exception e)
-                {
-                    completion.SetException(e);
-                }
-            };
-#pragma warning restore 1998
-
-            try
-            {
-                await _serviceClient.ExecuteQueryStringAsync(_tempFileUri, command.Code);
-
-                context.CancellationToken.Register(() =>
-                {
-
-                    _serviceClient.CancelQueryExecutionAsync(_tempFileUri)
-                        .Wait(TimeSpan.FromSeconds(10));
-
-                    completion.TrySetCanceled(context.CancellationToken);
-                });
-                await completion.Task;
-            }
-            catch (TaskCanceledException)
-            {
-                context.Display("Query cancelled.");
-            }
-            catch (OperationCanceledException)
-            {
-                context.Display("Query cancelled.");
-            }
-            finally
-            {
-                _queryCompletionHandler = null;
-                _queryMessageHandler = null;
-            }
-        }
-
-        private IEnumerable<IEnumerable<IEnumerable<(string name, object value)>>> GetEnumerableTables(ColumnInfo[] columnInfos, CellValue[][] rows)
-        {
-            var displayTable = new List<(string, object)[]>();
-            var columnNames = columnInfos.Select(info => info.ColumnName).ToArray();
-
-            SqlKernelUtils.AliasDuplicateColumnNames(columnNames);
-
-            foreach (CellValue[] row in rows)
-            {
-                var displayRow = new (string, object)[row.Length];
-
-                for (var colIndex = 0; colIndex < row.Length; colIndex++)
-                {
-                    object convertedValue = default;
-
-                    try
-                    {
-                        var columnInfo = columnInfos[colIndex];
-
-                        var expectedType = Type.GetType(columnInfo.DataType);
-
-                        if (TypeDescriptor.GetConverter(expectedType) is { } typeConverter)
-                        {
-                            if (typeConverter.CanConvertFrom(typeof(string)))
-                            {
-                                // TODO:fix handling target boolean type when the column is bit type with numeric value
-                                if ((expectedType == typeof(bool) || expectedType == typeof(bool?)) &&
-
-                                    decimal.TryParse(row[colIndex].DisplayValue, out var numericValue))
-                                {
-                                    convertedValue = numericValue != 0;
-                                }
-                                else
-                                {
-                                    convertedValue =
-                                        typeConverter.ConvertFromInvariantString(row[colIndex].DisplayValue);
-                                }
-                            }
-                        }
-                    }
-                    catch (Exception)
-                    {
-                        convertedValue = row[colIndex].DisplayValue;
-                    }
-
-                    displayRow[colIndex] = (columnNames[colIndex], convertedValue);
-                }
-
-                displayTable.Add(displayRow);
-            }
-
-            yield return displayTable;
-        }
-
-        public async Task HandleAsync(RequestCompletions command, KernelInvocationContext context)
-        {
-            if (!_intellisenseReady)
-            {
-                return;
-            }
-
-            var completionItems = await _serviceClient.ProvideCompletionItemsAsync(_tempFileUri, command);
-            context.Publish(new CompletionsProduced(completionItems, command));
+            return Type.GetType(typeName);
         }
 
         protected override ChooseKernelDirective CreateChooseKernelDirective() =>
